refactor(upgrade): change namespace for tests
