+++ packages/upgrade/tests/Tempest20/Fixtures/MigrateUpMigration.expected.php
@@ -1,26 +1,20 @@
 <?php
 
-namespace Tempest\Upgrade\Tests\Tempest2\Fixtures;
+namespace Tempest\Upgrade\Tests\Tempest20\Fixtures;
 
-use Tempest\Database\DatabaseMigration;
 use Tempest\Database\QueryStatement;
 use Tempest\Database\QueryStatements\CreateTableStatement;
 
-final class MigrateUpMigration implements DatabaseMigration
+final class MigrateUpMigration implements \Tempest\Database\MigratesUp
 {
     public string $name {
         get => '00-00-0000';
     }
 
-    public function up(): ?QueryStatement
+    public function up(): QueryStatement
     {
         return new CreateTableStatement('table')
             ->primary()
             ->datetime('createdAt');
-    }
-
-    public function down(): ?QueryStatement
-    {
-        return null;
     }
 }
